% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.4_method1.R
\name{area_continuous_paired}
\alias{area_continuous_paired}
\title{compare area and partial area for continuous data and paired design}
\usage{
area_continuous_paired(data1_no, data2_no, data1_yes, data2_yes, e1, e2, alpha)
}
\arguments{
\item{data1_no}{test results of undiseased individuals in test 1}

\item{data2_no}{test results of undiseased individuals in test 2}

\item{data1_yes}{test results of diseased individuals in test 1}

\item{data2_yes}{test results of diseased individuals in test 2}

\item{e1}{lower bound of FPR for the partial area}

\item{e2}{upper bound of FPR for the partial area}

\item{alpha}{significance level}
}
\value{
list containing the test statistics and whether the areas or partial areas are significantly different
example
data1_no<-c(136,286,281,23,200,146,220,96,100)
data1_yes<-c(140,1087,230,183,1256,700,16,800,253,740,126,153,283,90,303,193,76,1360,543)
data2_no<-c(60,126,100,40,253,46,70,17,27)
data2_yes<-c(323,671,350,156,1560,543,443,509,60,230,490,80,356,120,523,303,76,353,576)
area_continuous_paired(data1_no, data2_no, data1_yes, data2_yes, 0, 0.2, 0.05)
}
\description{
compare area and partial area for continuous data and paired design
}
