% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.4_basicfunctions.R
\name{area_estimation_continuous}
\alias{area_estimation_continuous}
\title{Estimate full area and partial area under ROC curve and their variances for continuous data}
\usage{
area_estimation_continuous(a, b, var_a, var_b, covar_ab, e1, e2)
}
\arguments{
\item{a}{binormal parameter a}

\item{b}{binormal parameter b}

\item{var_a}{estimation of variance of a}

\item{var_b}{estimation of variance of b}

\item{covar_ab}{estimation of covariance of binormal parameters a and b}

\item{e1}{lower bound of FPR for the partial area}

\item{e2}{upper bound of FPR for the partial area}
}
\description{
Estimate full area and partial area under ROC curve and their variances for continuous data
}
