% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.4_method1.R
\name{area_ordinal_paired}
\alias{area_ordinal_paired}
\title{compare area and partial area for ordinal data and paired design}
\usage{
area_ordinal_paired(data0, data1, e1, e2, alpha)
}
\arguments{
\item{data0}{a matrix representing paired test results for undiseased subjects}

\item{data1}{a matrix representing paired test results for diseased subjects}

\item{e1}{lower bound of FPR for the partial area}

\item{e2}{upper bound of FPR for the partial area}

\item{alpha}{significance level}
}
\value{
list containing the test statistics and whether the areas or partial areas are significantly different
example
data0<-matrix(c(36,20,8,6,0,0,3,3,1,0,0,1,0,4,0,0,0,2,3,2,2,1,2,5,2),nrow=5,byrow=TRUE)
data1<-matrix(c(1,2,2,2,1,0,0,0,0,2,0,0,0,2,0,0,0,0,2,0,0,0,1,8,39),nrow=5,byrow=TRUE)
area_ordinal_paired(data0, data1, 0, 0.2, 0.05)
}
\description{
compare area and partial area for ordinal data and paired design
}
