% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.4_method1.R
\name{area_ordinal_unpaired}
\alias{area_ordinal_unpaired}
\title{compare area and partial area for ordinal data and unpaired design}
\usage{
area_ordinal_unpaired(data1_no, data2_no, data1_yes, data2_yes, e1, e2, alpha)
}
\arguments{
\item{data1_no}{test results of undiseased individuals in test 1}

\item{data2_no}{test results of undiseased individuals in test 2}

\item{data1_yes}{test results of diseased individuals in test 1}

\item{data2_yes}{test results of diseased individuals in test 2}

\item{e1}{lower bound of FPR for the partial area}

\item{e2}{upper bound of FPR for the partial area}

\item{alpha}{significance level}
}
\value{
list containing the test statistics and whether the areas or partial areas are significantly different
example
data1_no<-c(38,25,15,19,4)
data1_yes<-c(1,2,3,14,42)
data2_no<-c(70,7,5,7,12)
data2_yes<-c(8,2,2,2,48)
area_ordinal_unpaired(data1_no, data2_no, data1_yes, data2_yes, 0, 0.2, 0.05)
}
\description{
compare area and partial area for ordinal data and unpaired design
}
