% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.4_method6.R
\name{cluster_area_compare}
\alias{cluster_area_compare}
\title{Input two matrices, each corresponding to a test.Each matrix consists of two rows per group.One group corresponds to one cluster.
The first row of each group corresponds to the detection results of diseased units in that cluster
The second row of each group corresponds to the detection results of undiseased units in that cluster
All are categorized into classes 1, 2,...,K
Note that two tests should be implemented on same subjects}
\usage{
cluster_area_compare(data1, data2, alpha)
}
\arguments{
\item{data1}{test result of clustered data for test 1}

\item{data2}{test result of clustered data for test 2}

\item{alpha}{significance level}
}
\value{
a list containing the test statistic and a conclusion on whether the areas are significantly different
example
data1 <- matrix(c(38,25,15,19,4,1,2,3,14,42,70,7,5,7,12,8,2,2,2,48), nrow=4, byrow=TRUE)
data2 <- matrix(c(38,25,19,15,4,3,2,1,14,42,70,5,7,7,12,2,2,2,8,48), nrow=4, byrow=TRUE)
cluster_area_compare(data1, data2, 0.05)
}
\description{
Input two matrices, each corresponding to a test.Each matrix consists of two rows per group.One group corresponds to one cluster.
The first row of each group corresponds to the detection results of diseased units in that cluster
The second row of each group corresponds to the detection results of undiseased units in that cluster
All are categorized into classes 1, 2,...,K
Note that two tests should be implemented on same subjects
}
