% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1.2.R
\name{clustered_binary_Sp}
\alias{clustered_binary_Sp}
\title{compare specificities for clustered binary data}
\usage{
clustered_binary_Sp(data1, data2, alpha)
}
\arguments{
\item{data1}{binary test results for test 1. It contains two rows.
The first row is the number of true negative elements in each cluster
The second row is the number of negative elements on gold standard in each cluster}

\item{data2}{binary test results for test 2. It contains two rows.
The first row is the number of true negative elements in each cluster
The second row is the number of negative elements on gold standard in each cluster}

\item{alpha}{significance level}
}
\value{
list containing the test statistic and whether specificities are significantly different
example
data1<-matrix(c(0,2,2,0,2,2,1,1,0,1,2,0,1,2,1,1,1,1,1,0,1,2,1,2,0,1,2,2,1,2,2,1,1,1,1,2,1,3,2,1,1,1,2,2,2,1,2,2,2,1),nrow=2,byrow=TRUE)
data2<-matrix(c(1,2,2,1,2,2,1,1,1,1,2,0,2,2,1,1,1,2,1,0,1,2,2,2,0,1,2,2,1,2,2,1,1,1,1,2,1,3,2,1,1,1,2,2,2,1,2,2,2,1),nrow=2,byrow=TRUE)
clustered_binary_Sp(data1, data2, 0.05)
}
\description{
compare specificities for clustered binary data
}
