% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.4_method5.R
\name{compare_partial_He}
\alias{compare_partial_He}
\title{He and Escobar's method to compare partial area for continuous paired data}
\usage{
compare_partial_He(data1_no, data1_yes, data2_no, data2_yes, e1, e2, alpha)
}
\arguments{
\item{data1_no}{continuous test result of test 1 for undiseased subjects}

\item{data1_yes}{continuous test result of test 1 for diseased subjects}

\item{data2_no}{continuous test result of test 2 for undiseased subjects}

\item{data2_yes}{continuous test result of test 2 for diseased subjects}

\item{e1}{lower bound of FPR for partial area}

\item{e2}{upper bound of FPR for partial area}

\item{alpha}{significance level}
}
\value{
list containing the test statistic and whether the partial areas are significantly different
example
data1_no <- c(136,286,281,23,200,146,220,96,100)
data1_yes <- c(140,1087,230,183,1256,700,16,800,253,740,126,153,283,90,303,193,76,1360,543)
data2_no <- c(60,126,100,40,253,46,70,17,27)
data2_yes <- c(323,671,350,156,1560,543,443,509,60,230,490,80,356,120,523,303,76,353,576)
compare_partial_He(data1_no, data1_yes, data2_no, data2_yes, 0, 1, 0.05)
}
\description{
He and Escobar's method to compare partial area for continuous paired data
}
