% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.3.R
\name{differ_ordinal_unpaired}
\alias{differ_ordinal_unpaired}
\title{Determining the Range of FPRs for which TPRs Differ for ordinal unpaired data}
\usage{
differ_ordinal_unpaired(data1_no, data2_no, data1_yes, data2_yes, alpha)
}
\arguments{
\item{data1_no}{test results of undiseased individuals in test 1}

\item{data2_no}{test results of undiseased individuals in test 2}

\item{data1_yes}{test results of diseased individuals in test 1}

\item{data2_yes}{test results of diseased individuals in test 2}

\item{alpha}{significance level}
}
\value{
list containing the range of FPRs for which TPRs differ
example
data1_no<-c(38,25,15,19,4)
data1_yes<-c(1,2,3,14,42)
data2_no<-c(70,7,5,7,12)
data2_yes<-c(8,2,2,2,48)
differ_ordinal_unpaired(data1_no, data2_no, data1_yes, data2_yes, 0.05)
}
\description{
Determining the Range of FPRs for which TPRs Differ for ordinal unpaired data
}
