% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.4_method4.R
\name{generate}
\alias{generate}
\title{transform ordinal test results into a vector that contains test result of each individual.It is not important.}
\usage{
generate(data)
}
\arguments{
\item{data}{ordinal test results}
}
\description{
transform ordinal test results into a vector that contains test result of each individual.It is not important.
}
