% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.1_basicfunctions.R
\name{params_estimate_continuous_paired}
\alias{params_estimate_continuous_paired}
\title{estimate the binormal parameters for paired continuous data,as well as their variances and covariances
这里默认data1_no,data1_yes,data2_no,data2_yes是原始数据,需要进行Box_Cox transformation.
using Zou's method to estimate parameters for continuous data under binormal assumption
test 1 and test 2 should be implemented on the same individuals}
\usage{
params_estimate_continuous_paired(data1_no, data1_yes, data2_no, data2_yes)
}
\arguments{
\item{data1_no}{continuous test result from undiseased individuals using test 1}

\item{data1_yes}{continuous test result from diseased individuals using test 1}

\item{data2_no}{continuous test result from undiseased individuals using test 2}

\item{data2_yes}{continuous test result from diseased individuals using test 2}
}
\value{
the estimated values of the parameters for two Box-Cox transformations,the corresponding binormal parameters
and the estimated covariance matrix
#example
#data1_no<-c(136,286,281,23,200,146,220,96,100)
#data1_yes<-c(140,1087,230,183,1256,700,16,800,253,740,126,153,283,90,303,193,76,1360,543)
#data2_no<-c(60,126,100,40,253,46,70,17,27)
#data2_yes<-c(323,671,350,156,1560,543,443,509,60,230,490,80,356,120,523,303,76,353,576)
#params_estimate_continuous_paired(data1_no,data1_yes,data2_no,data2_yes)
}
\description{
estimate the binormal parameters for paired continuous data,as well as their variances and covariances
这里默认data1_no,data1_yes,data2_no,data2_yes是原始数据,需要进行Box_Cox transformation.
using Zou's method to estimate parameters for continuous data under binormal assumption
test 1 and test 2 should be implemented on the same individuals
}
