% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.1_basicfunctions.R
\name{params_estimate_ordinal_paired}
\alias{params_estimate_ordinal_paired}
\title{estimate the binormal parameters for paired ordinal data,as well as their variances and covariances}
\usage{
params_estimate_ordinal_paired(data0, data1)
}
\arguments{
\item{data0}{a matrix representing ordinal test result of undiseased individuals}

\item{data1}{a matrix representing ordinal test result of diseased individuals}
}
\value{
estimation of decision thresholds,binormal parameters and their covariance matrix
#example
#data0<-matrix(c(36,20,8,6,0,0,3,3,1,0,0,1,0,4,0,0,0,2,3,2,2,1,2,5,2),nrow=5,byrow=TRUE)
#data1<-matrix(c(1,2,2,2,1,0,0,0,0,2,0,0,0,2,0,0,0,0,2,0,0,0,1,8,39),nrow=5,byrow=TRUE)
#params_estimate_ordinal_paired(data0,data1)
}
\description{
estimate the binormal parameters for paired ordinal data,as well as their variances and covariances
}
