% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.1_basicfunctions.R
\name{params_estimate_ordinal_unpaired}
\alias{params_estimate_ordinal_unpaired}
\title{estimate the binormal parameters for unpaired ordinal data,as well as their variances and covariances}
\usage{
params_estimate_ordinal_unpaired(data0, data1)
}
\arguments{
\item{data0}{ordinal test result from undiseased individuals}

\item{data1}{ordinal test result from diseased individuals}
}
\value{
estimation of decision thresholds,binormal parameters and their covariance matrix
#example
#data0<-c(38,25,15,19,4)
#data1<-c(1,2,3,14,42)
#params_estimate_ordinal_unpaired(data0,data1)
}
\description{
estimate the binormal parameters for unpaired ordinal data,as well as their variances and covariances
}
