% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.1.3.R
\name{PPV_Leisenring}
\alias{PPV_Leisenring}
\title{compare PPVs using Leisenring's method}
\usage{
PPV_Leisenring(data1, data2, alpha)
}
\arguments{
\item{data1}{data from nondiseased individuals in the form of table 5.2 of the textbook}

\item{data2}{data from diseased individuals in the form of table 5.2 of the textbook}

\item{alpha}{significance level}
}
\value{
list containing the test statistic, p-value, and whether PPVs are significantly different
example
data1<-matrix(c(12,7,11,71),nrow=2,byrow=TRUE)
data2<-matrix(c(49,1,7,5),nrow=2,byrow=TRUE)
PPV_Leisenring(data1, data2, 0.05)
}
\description{
compare PPVs using Leisenring's method
}
