% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.4_method4.R
\name{Psi}
\alias{Psi}
\title{this function has the same definition as the function Psi in the book.It is not important.}
\usage{
Psi(m, n)
}
\arguments{
\item{m}{the same definition as the function Psi in the book.}

\item{n}{the same definition as the function Psi in the book.}
}
\description{
this function has the same definition as the function Psi in the book.It is not important.
}
