% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.4_method4.R
\name{roc_compare_Delong}
\alias{roc_compare_Delong}
\title{using Delong's method to determine whether the areas are significantly different for ordinal paired data}
\usage{
roc_compare_Delong(data1_yes, data1_no, data2_yes, data2_no, alpha)
}
\arguments{
\item{data1_yes}{ordinal test results of test 1 for diseased subjects}

\item{data1_no}{ordinal test results of test 1 for undiseased subjects}

\item{data2_yes}{ordinal test results of test 2 for diseased subjects}

\item{data2_no}{ordinal test results of test 2 for undiseased subjects}

\item{alpha}{significance level}
}
\value{
list containing the test statistic and whether the areas are significantly different
example
data1_no <- c(38, 25, 15, 19, 4)
data1_yes <- c(1, 2, 3, 14, 42)
data2_no <- c(70, 7, 5, 7, 12)
data2_yes <- c(8, 2, 2, 2, 48)
roc_compare_Delong(data1_yes, data1_no, data2_yes, data2_no, 0.05)
}
\description{
using Delong's method to determine whether the areas are significantly different for ordinal paired data
}
