% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.1_binormalparameters.R
\name{roctest_continuous_paired}
\alias{roctest_continuous_paired}
\title{using Zou's method to determine whether binormal parameters are equal for continuous paired data
test 1 and test 2 should be implemented on the same subjects}
\usage{
roctest_continuous_paired(data1_no, data1_yes, data2_no, data2_yes, alpha)
}
\arguments{
\item{data1_no}{test results of test 1 for subjects without disease}

\item{data1_yes}{test results of test 1 for subjects with disease}

\item{data2_no}{test results of test 2 for subjects without disease}

\item{data2_yes}{test results of test 2 for subjects with disease}

\item{alpha}{significance level}
}
\value{
list containing the test statistic and whether the binormal parameters are equal
example
data1_no<-c(136,286,281,23,200,146,220,96,100)
data1_yes<-c(140,1087,230,183,1256,700,16,800,253,740,126,153,283,90,303,193,76,1360,543)
data2_no<-c(60,126,100,40,253,46,70,17,27)
data2_yes<-c(323,671,350,156,1560,543,443,509,60,230,490,80,356,120,523,303,76,353,576)
roctest_continuous_paired(data1_no, data1_yes, data2_no, data2_yes, 0.05)
}
\description{
using Zou's method to determine whether binormal parameters are equal for continuous paired data
test 1 and test 2 should be implemented on the same subjects
}
