% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.1_binormalparameters.R
\name{roctest_ordinal_paired}
\alias{roctest_ordinal_paired}
\title{test whether binormal parameters are equal for paired ordinal data}
\usage{
roctest_ordinal_paired(data0, data1, alpha)
}
\arguments{
\item{data0}{a matrix representing paired test results of undiseased individuals}

\item{data1}{a matrix representing paired test results of diseased individuals}

\item{alpha}{significance level}
}
\value{
list containing the test statistic and whether the binormal parameters are equal
example
data0<-matrix(c(36,20,8,6,0,0,3,3,1,0,0,1,0,4,0,0,0,2,3,2,2,1,2,5,2),nrow=5,byrow=TRUE)
data1<-matrix(c(1,2,2,2,1,0,0,0,0,2,0,0,0,2,0,0,0,0,2,0,0,0,1,8,39),nrow=5,byrow=TRUE)
roctest_ordinal_paired(data0, data1, 0.05)
}
\description{
test whether binormal parameters are equal for paired ordinal data
}
