% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.1_binormalparameters.R
\name{roctest_ordinal_unpaired}
\alias{roctest_ordinal_unpaired}
\title{test whether binormal parameters are equal for unpaired ordinal data}
\usage{
roctest_ordinal_unpaired(data1_no, data2_no, data1_yes, data2_yes, alpha)
}
\arguments{
\item{data1_no}{test results of undiseased individuals in test 1}

\item{data2_no}{test results of undiseased individuals in test 2}

\item{data1_yes}{test results of diseased individuals in test 1}

\item{data2_yes}{test results of diseased individuals in test 2}

\item{alpha}{significance level}
}
\value{
list containing the test statistic and whether the binormal parameters are equal
example
data1_no<-c(38,25,15,19,4)
data1_yes<-c(1,2,3,14,42)
data2_no<-c(70,7,5,7,12)
data2_yes<-c(8,2,2,2,48)
roctest_ordinal_unpaired(data1_no, data2_no, data1_yes, data2_yes, 0.05)
}
\description{
test whether binormal parameters are equal for unpaired ordinal data
}
