% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.2_parametric.R
\name{TPRtest_ordinal_paired}
\alias{TPRtest_ordinal_paired}
\title{function to compare sensitivities at a particular point for ordinal data and paired design}
\usage{
TPRtest_ordinal_paired(data0, data1, alpha, e)
}
\arguments{
\item{data0}{a matrix representing paired test results of undiseased individuals}

\item{data1}{a matrix representing paired test results of diseased individuals}

\item{alpha}{significance level}

\item{e}{FPR rate of this point}
}
\value{
list containing the test statistic and whether the sensitivities (TPRs) are significantly different
#example
data0<-matrix(c(36,20,8,6,0,0,3,3,1,0,0,1,0,4,0,0,0,2,3,2,2,1,2,5,2),nrow=5,byrow=TRUE)
data1<-matrix(c(1,2,2,2,1,0,0,0,0,2,0,0,0,2,0,0,0,0,2,0,0,0,1,8,39),nrow=5,byrow=TRUE)
TPRtest_ordinal_paired(data0, data1, 0.05, 0.1)
}
\description{
function to compare sensitivities at a particular point for ordinal data and paired design
}
