% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.2.2_parametric.R
\name{TPRtest_ordinal_unpaired}
\alias{TPRtest_ordinal_unpaired}
\title{function to compare sensitivities at a particular point for ordinal data and unpaired design}
\usage{
TPRtest_ordinal_unpaired(data1_no, data2_no, data1_yes, data2_yes, alpha, e)
}
\arguments{
\item{data1_no}{test results of undiseased individuals in test 1}

\item{data2_no}{test results of undiseased individuals in test 2}

\item{data1_yes}{test results of diseased individuals in test 1}

\item{data2_yes}{test results of diseased individuals in test 2}

\item{alpha}{significance level}

\item{e}{FPR rate of this point}
}
\value{
list containing the test statistic and whether the sensitivities (TPRs) are significantly different
#example
data1_no<-c(38,25,15,19,4)
data1_yes<-c(1,2,3,14,42)
data2_no<-c(70,7,5,7,12)
data2_yes<-c(8,2,2,2,48)
TPRtest_ordinal_unpaired(data1_no, data2_no, data1_yes, data2_yes, 0.05, 0.1)
}
\description{
function to compare sensitivities at a particular point for ordinal data and unpaired design
}
